Refactor: Separate Pydantic models into individual files
I've restructured the Pydantic database models (schemas) from a single file (`infrastructure/database/schemas.py`) into individual files, each housing one model, located under the `app/models/` directory.

Key changes include:
- I moved PersonDB, CaseManagementDB, RawEventDB, CompanyProfileDB, BeneficialOwnerDB, StoredEventMetaData, StoredEventDB, and RequiredDocumentDB to their own files in `app/models/`.
- I created `app/models/__init__.py` to provide easy access to these models.
- I updated all import statements throughout the project (including source code, tests, and documentation) to reflect the new model locations.
- I cleaned up the old `infrastructure/database/schemas.py` file, retaining only essential imports like `AddressData`.
- I removed unused `app/models/commands/` and `app/models/events/` directories as you confirmed.
- I performed extensive linting and test fixing to ensure the stability of the application after these changes. This involved correcting import paths, fixing issues related to database connection handling (`get_db` async generator), updating `unittest.mock.patch` targets, and resolving various other errors uncovered during testing.

This refactoring improves modularity, making the codebase cleaner, more scalable, and easier to maintain.
