+++ case_management_service/app/models/stored_event_meta_data.py
@@ -1,0 +1,8 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class StoredEventMetaData(BaseModel):
+    correlation_id: Optional[str] = None
+    causation_id: Optional[str] = None
